Fix bug keeping page from updating the image if left open over the hour.
The page would only read the time at the initial page load and set the
corresponding image/message text, but would not update when the hour value
changed, as that was being changed by the updateClock function, and the
code setting the image was pulling from new (Date).getHours and there was
nothing re-setting that variable's value. Added code in the updateClock
function that when minutes and seconds == 00 it called the function to
set the Image/Message. Also added a re-set for the time variable within
the setImageMssg function, so that the correct hour was being read
before executing the if/then statements.